feat: versions of `ind₁`/`coind₁` defined as finsupp/pi `ind₁`/`coind₁` are defined as the base change of finsupp/pi quotiented out by the trivial relation. This is because they are abbrevs of the general construction from mathlib.
Instead of redefining them as `G →₀ A`/`G → A` with the `G`-action on the domain, which would break
the defeq with the general construction, we provide `ind₁AsFinsupp`/`coind₁AsPi`, a version of
`ind₁`/`coind₁` that's actually defined as `G →₀ A`/`G → A`.

`ind₁AsFinsupp`/`coind₁AsPi` are not bundled as functors because they should only be used for
pointwise computations.

Co-authored-by: Arend Mellendijk
Co-authored-by: Michał Mrugała <kiolterino@gmail.com>