chore: clean after #45 (#47)
Edison's PR wasn't using the `↓` notation for restriction because he was in a universe-polymorphic file and the `↓` notation was universe-monomorphic. I could either make the latter file universe-monomorphic, or the notation universe-polymorphic. I did the latter because it was not exactly trivial and since it might soon be included in mathlib.