Periodicity of Tate cohomology (#86)
Rename definitions in the `up`/`down` API to be in lowerCamelCase and
lemmas to better follow the naming convention.

Improve the Tate cohomology and `up`/`down` APIs by syntactically
generalising declarations: Instead of having them be about `n + 1 :
Nat`, we make them be about `n : Nat` + `NeZero n`. Among other places,
this reduces friction in the proof of periodicity.

Restore the declarations from #34 that were accidentally deleted in #45.
See
https://github.com/kbuzzard/ClassFieldTheory/pull/85#issuecomment-3449775696
for more context.

Put all of this together into a proof that Tate cohomology is periodic,
as well as a computation of the even and odd Tate cohomology of an
arbitrary finite cyclic group. These were the final few sorries in
`Cohomology.FiniteCyclic.UpDown`.

Closes #85, which is proving a subset of those sorries but redoing the
work from #34.